Updated: styles
This in production, would not be a single monolithic file, I completely
understand how work is split among developers these days. So know that
this was intentional.

My reasoning was:

1. I thought a single file would be easier on the reviewer, given the
contect of this being a "try out" project and not meant for production.

2. A side benefit is that I wasn't happy with how the components and their
styles were arranged. Normally I would have built, say the table component
on its lonesome, included the styles alongside and imported that component
in the two routes its required. Seeing that it was baked into each route
styles wouldn't share properly without duplication (then increasing blocking
the main thread... I promise I remember its only a test app). The answer
is then to share it in a stylesheet that both route components share.

I also have not included the fonts in this repo. I was not at all
confident that the fonts I had procured were above board legally to
include in the repo.

**Upon request, I am more than happy to prove I understand how fonts
ought to be loaded.** Again, I simply wanted to reduce the opportunity
for a DMCA take down on a test application repo.